Make the midi library channel aware
Behold curious tinkerer! I've resurrected the channel support in the
midi library, now you can register the channel number with the
SetChannel() and SetChannelCall() methods:

SetChannel()     -> Register a fixed channel;
SetChannelCall() -> Register a callback that returns the selected midi
channel, it is useful when the channel changes dynamically through
external inputs.

I know, it's a little confusing. Maybe I'll change it in the future...